start extracting utility class
